DinhDuong/src/main/webapp/app/entities/ho-tro/delete/ho-tro-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { HoTroService } from '../service/ho-tro.service';

import { HoTroDeleteDialogComponent } from './ho-tro-delete-dialog.component';

describe('HoTro Management Delete Component', () => {
  let comp: HoTroDeleteDialogComponent;
  let fixture: ComponentFixture<HoTroDeleteDialogComponent>;
  let service: HoTroService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [HoTroDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(HoTroDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(HoTroDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(HoTroService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
